Reordering and deleting taps due to network shutdown.
Due to recent changes in regulation of centralized exchanges, we have decided to shutdown seed nodes of Talleo network on September 30, 2025. There has not been any trading on exchanges were TLO was listed on since June 20, 2024 and that makes it very hard to continue as value of TLO hasn’t gone up to levels that would make it sustainable. The network will continue running as long as there is third-party mining pools or solo miners, but all the official websites like webwallet and the social media bots will be taken offline. No new nodes can join the network after the seed nodes are shut down.